+++ python/pathway/internals/datasink.py
@@ -5,14 +5,29 @@ from __future__ import annotations
 from abc import ABC
 from collections.abc import Callable
 from dataclasses import dataclass
+from typing import TYPE_CHECKING, Iterable
 
 from pathway.internals import api
+from pathway.internals.expression import ColumnReference
+
+if TYPE_CHECKING:
+    from pathway.internals.table import Table
 
 
 class DataSink(ABC):
     @property
     def name(self) -> str:
         return type(self).__qualname__.lower().removesuffix("datasink")
+
+    def check_sort_by_columns(self, table: Table):
+        sort_by = getattr(self, "sort_by", None)
+        if sort_by is None:
+            return
+        for column in sort_by:
+            if column._table != table:
+                raise ValueError(
+                    f"The column {column} doesn't belong to the target table {table}"
+                )
 
 
 @dataclass(frozen=True)
@@ -21,10 +36,20 @@ class GenericDataSink(DataSink):
     dataformat: api.DataFormat
     datasink_name: str
     unique_name: str | None
+    sort_by: Iterable[ColumnReference] | None = None
 
     @property
     def name(self) -> str:
         return self.datasink_name
+
+    @property
+    def sort_by_indices(self):
+        if self.sort_by is None:
+            return None
+        column_index: dict[str, int] = {}
+        for index, column in enumerate(self.dataformat.value_fields):
+            column_index[column.name] = index
+        return [column_index[column.name] for column in self.sort_by]
 
 
 @dataclass(frozen=True, kw_only=True)
@@ -35,6 +60,15 @@ class CallbackDataSink(DataSink):
     skip_persisted_batch: bool
     skip_errors: bool
     unique_name: str | None
+    sort_by: Iterable[ColumnReference] | None = None
+
+    def sort_by_indices(self, table: Table):
+        if self.sort_by is None:
+            return None
+        column_index: dict[str, int] = {}
+        for index, column in enumerate(table._columns):
+            column_index[column] = index
+        return [column_index[column.name] for column in self.sort_by]
 
 
 @dataclass(frozen=True)