+++ python/pathway/internals/table_subscription.py
@@ -2,10 +2,11 @@
 
 from __future__ import annotations
 
-from typing import Any, Protocol
+from typing import Any, Iterable, Protocol
 
 from pathway.internals import datasink
 from pathway.internals.api import Pointer
+from pathway.internals.expression import ColumnReference
 from pathway.internals.table_io import table_to_datasink
 
 
@@ -84,6 +85,7 @@ def subscribe(
     on_end: OnFinishCallback = lambda: None,
     skip_errors: bool = True,
     name: str | None = None,
+    sort_by: Iterable[ColumnReference] | None = None,
 ) -> None:
     """
     Calls a callback function on_change on every change happening in table. This method
@@ -106,6 +108,9 @@ def subscribe(
         skip_errors: whether to skip rows containing errors
         name: A unique name for the connector. If provided, this name will be used in
             logs and monitoring dashboards.
+        sort_by: If specified, the output will be sorted in ascending order based on the
+            values of the given columns within each minibatch. When multiple columns are provided,
+            the corresponding value tuples will be compared lexicographically.
     Returns:
         None
     """
@@ -146,5 +151,6 @@ def subscribe(
             skip_persisted_batch=skip_persisted_batch,
             skip_errors=skip_errors,
             unique_name=name,
+            sort_by=sort_by,
         ),
     )