+++ python/pathway/io/deltalake/__init__.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from os import PathLike, fspath
-from typing import Any
+from typing import Any, Iterable
 
 from pathway.internals import api, datasink, datasource
 from pathway.internals._io_helpers import (
@@ -12,6 +12,7 @@ from pathway.internals._io_helpers import (
     is_s3_path,
 )
 from pathway.internals.config import _check_entitlements
+from pathway.internals.expression import ColumnReference
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.schema import Schema
 from pathway.internals.table import Table
@@ -190,6 +191,7 @@ def write(
     ) = None,
     min_commit_frequency: int | None = 60_000,
     name: str | None = None,
+    sort_by: Iterable[ColumnReference] | None = None,
 ) -> None:
     """
     Writes the stream of changes from ``table`` into `Delta Lake <https://delta.io/>_` data
@@ -225,6 +227,9 @@ def write(
             operations afterwards.
         name: A unique name for the connector. If provided, this name will be used in
             logs and monitoring dashboards.
+        sort_by: If specified, the output will be sorted in ascending order based on the
+            values of the given columns within each minibatch. When multiple columns are provided,
+            the corresponding value tuples will be compared lexicographically.
 
     Returns:
         None
@@ -287,5 +292,6 @@ def write(
             data_format,
             datasink_name="deltalake",
             unique_name=name,
+            sort_by=sort_by,
         )
     )