+++ python/pathway/io/elasticsearch/__init__.py
@@ -2,8 +2,11 @@
 
 from __future__ import annotations
 
+from typing import Iterable
+
 from pathway.internals import api, datasink
 from pathway.internals._io_helpers import _format_output_value_fields
+from pathway.internals.expression import ColumnReference
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.table import Table
 from pathway.internals.trace import trace_user_frame
@@ -56,6 +59,7 @@ def write(
     index_name: str,
     *,
     name: str | None = None,
+    sort_by: Iterable[ColumnReference] | None = None,
 ) -> None:
     """Write a table to a given index in ElasticSearch.
 
@@ -66,6 +70,9 @@ def write(
         index_name: name of the index, which gets the docs.
         name: A unique name for the connector. If provided, this name will be used in
             logs and monitoring dashboards.
+        sort_by: If specified, the output will be sorted in ascending order based on the
+            values of the given columns within each minibatch. When multiple columns are provided,
+            the corresponding value tuples will be compared lexicographically.
 
     Returns:
         None
@@ -123,5 +130,6 @@ def write(
             data_format,
             datasink_name="elasticsearch",
             unique_name=name,
+            sort_by=sort_by,
         )
     )