+++ python/pathway/io/fs/__init__.py
@@ -4,10 +4,11 @@ from __future__ import annotations
 
 import warnings
 from os import PathLike, fspath
-from typing import Any
+from typing import Any, Iterable
 
 from pathway.internals import Schema, api, datasink, datasource
 from pathway.internals._io_helpers import _format_output_value_fields
+from pathway.internals.expression import ColumnReference
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.table import Table
 from pathway.internals.table_io import table_from_datasource
@@ -261,6 +262,7 @@ def write(
     format: str,
     *,
     name: str | None = None,
+    sort_by: Iterable[ColumnReference] | None = None,
 ) -> None:
     """Writes ``table``'s stream of updates to a file in the given format.
 
@@ -271,6 +273,9 @@ def write(
             formats: "json" and "csv".
         name: A unique name for the connector. If provided, this name will be used in
             logs and monitoring dashboards.
+        sort_by: If specified, the output will be sorted in ascending order based on the
+            values of the given columns within each minibatch. When multiple columns are provided,
+            the corresponding value tuples will be compared lexicographically.
 
     Returns:
         None
@@ -353,6 +358,10 @@ def write(
 
     table.to(
         datasink.GenericDataSink(
-            data_storage, data_format, datasink_name="fs", unique_name=name
+            data_storage,
+            data_format,
+            datasink_name="fs",
+            unique_name=name,
+            sort_by=sort_by,
         )
     )