+++ python/pathway/io/postgres/__init__.py
@@ -2,10 +2,11 @@
 
 from __future__ import annotations
 
-from typing import Literal
+from typing import Iterable, Literal
 
 from pathway.internals import api, datasink
 from pathway.internals._io_helpers import _format_output_value_fields
+from pathway.internals.expression import ColumnReference
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.table import Table
 from pathway.internals.trace import trace_user_frame
@@ -37,6 +38,7 @@ def write(
     max_batch_size: int | None = None,
     init_mode: Literal["default", "create_if_not_exists", "replace"] = "default",
     name: str | None = None,
+    sort_by: Iterable[ColumnReference] | None = None,
 ) -> None:
     """Writes ``table``'s stream of updates to a postgres table.
 
@@ -55,6 +57,9 @@ def write(
             "replace": Initializes the SQL writer by replacing any existing table.
         name: A unique name for the connector. If provided, this name will be used in
             logs and monitoring dashboards.
+        sort_by: If specified, the output will be sorted in ascending order based on the
+            values of the given columns within each minibatch. When multiple columns are provided,
+            the corresponding value tuples will be compared lexicographically.
 
     Returns:
         None
@@ -134,6 +139,7 @@ def write(
             data_format,
             datasink_name="postgres.sink",
             unique_name=name,
+            sort_by=sort_by,
         )
     )
 
@@ -148,6 +154,7 @@ def write_snapshot(
     max_batch_size: int | None = None,
     init_mode: Literal["default", "create_if_not_exists", "replace"] = "default",
     name: str | None = None,
+    sort_by: Iterable[ColumnReference] | None = None,
 ) -> None:
     """Maintains a snapshot of a table within a Postgres table.
 
@@ -166,6 +173,9 @@ def write_snapshot(
             "replace": Initializes the SQL writer by replacing any existing table.
         name: A unique name for the connector. If provided, this name will be used in
             logs and monitoring dashboards.
+        sort_by: If specified, the output will be sorted in ascending order based on the
+            values of the given columns within each minibatch. When multiple columns are provided,
+            the corresponding value tuples will be compared lexicographically.
 
     Returns:
         None
@@ -231,5 +241,6 @@ def write_snapshot(
             data_format,
             datasink_name="postgres.snapshot",
             unique_name=name,
+            sort_by=sort_by,
         )
     )